Add `show_with_token`
`show_with_token` fetches the FAQ whose `id` was provided in `params`,
and only succeeds if the `token` and `Referer` header match the user's
`token` and `website_domain` respectively. If no FAQ matches the `id`,
we return a `404 Not Found` response. If the FAQ is found but either the
`token` or `Referer` doesn't match, we return a `401 Unauthorized`.